refine: centralise interfaces
