--- examples/aws-lambda/src/mock-transmitter/types.ts
@@ -1,21 +1,0 @@
-export interface SETVerificationRequest {
-  stream_id: string
-  state: string | undefined
-}
-
-export interface SET {
-  iss: string
-  aud: string
-  iat: number
-  jti: string
-  events: Record<
-    string,
-    {
-      state?: string
-    }
-  >
-  sub_id: {
-    format: 'opaque'
-    id: string
-  }
-}
